+++ Source/Memory/InlinePoolAllocator.cpp
@@ -12,7 +12,7 @@ namespace
     class VectorGrowthTest final
     {
     public:
-        float operator()(const char *name, auto &ls, auto &lf) const
+        float operator()(const char* name, auto& ls, auto& lf) const
         {
             using namespace HE;
 
@@ -20,7 +20,7 @@ namespace
             Time::TDuration stdTime;
 
             constexpr int testIterations = 8192;
-            constexpr size_t vectorSize = 128;
+            constexpr int vectorSize = 128;
 
             {
                 Time::ScopedTime measure(inlineTime);
@@ -74,14 +74,14 @@ namespace
 
 void HE::InlinePoolAllocatorTest::Prepare()
 {
-    AddTest("Zero-sized Allocation & Deallocation", [](auto &) {
+    AddTest("Zero-sized Allocation & Deallocation", [](auto&) {
         InlinePoolAllocator<int, 16> allocator;
 
         auto ptr = allocator.allocate(0);
         allocator.deallocate(ptr, 0);
     });
 
-    auto allocDeallocTest = [](auto &inlineAlloc, auto &ls, auto &lf) -> float {
+    auto allocDeallocTest = [](auto& inlineAlloc, auto& ls, auto& lf) -> float {
         constexpr int testCount = 500;
         constexpr int loopLength = 1024;
 
@@ -124,7 +124,7 @@ void HE::InlinePoolAllocatorTest::Prepare()
         auto rate = inlineTimeSec / stdTimeSec;
 
 #ifdef PROFILE_ENABLED
-        auto &mmgr = MemoryManager::GetInstance();
+        auto& mmgr = MemoryManager::GetInstance();
         auto stat = mmgr.GetAllocatorStat(inlineAlloc.GetID());
 
         ls << "Performance: " << inlineTimeSec << " msec vs STL: " << stdTimeSec
@@ -139,7 +139,7 @@ void HE::InlinePoolAllocatorTest::Prepare()
         return rate;
     };
 
-    AddTest("Perf. & Stress, Size(1)", [&, this](auto &ls) {
+    AddTest("Perf. & Stress, Size(1)", [&, this](auto& ls) {
         constexpr int inlineSize = 1;
         InlinePoolAllocator<int, inlineSize> inlineAlloc;
 
@@ -151,7 +151,7 @@ void HE::InlinePoolAllocatorTest::Prepare()
         }
     });
 
-    AddTest("Perf. & Stress, Size(16)", [&, this](auto &ls) {
+    AddTest("Perf. & Stress, Size(16)", [&, this](auto& ls) {
         constexpr int inlineSize = 16;
         InlinePoolAllocator<int, inlineSize> inlineAlloc;
 
@@ -163,7 +163,7 @@ void HE::InlinePoolAllocatorTest::Prepare()
         }
     });
 
-    AddTest("Perf. & Stress, Size(32)", [&, this](auto &ls) {
+    AddTest("Perf. & Stress, Size(32)", [&, this](auto& ls) {
         constexpr int inlineSize = 32;
         InlinePoolAllocator<int, inlineSize> inlineAlloc;
 
@@ -175,7 +175,7 @@ void HE::InlinePoolAllocatorTest::Prepare()
         }
     });
 
-    AddTest("Perf. & Stress, Size(64)", [&, this](auto &ls) {
+    AddTest("Perf. & Stress, Size(64)", [&, this](auto& ls) {
         constexpr int inlineSize = 64;
         InlinePoolAllocator<int, inlineSize> inlineAlloc;
 
@@ -187,7 +187,7 @@ void HE::InlinePoolAllocatorTest::Prepare()
         }
     });
 
-    AddTest("Perf. & Stress, Size(128)", [&, this](auto &ls) {
+    AddTest("Perf. & Stress, Size(128)", [&, this](auto& ls) {
         constexpr int inlineSize = 128;
         InlinePoolAllocator<int, inlineSize> inlineAlloc;
 
@@ -199,7 +199,7 @@ void HE::InlinePoolAllocatorTest::Prepare()
         }
     });
 
-    AddTest("Perf. & Stress, Size(256)", [&, this](auto &ls) {
+    AddTest("Perf. & Stress, Size(256)", [&, this](auto& ls) {
         constexpr int inlineSize = 256;
         InlinePoolAllocator<int, inlineSize> inlineAlloc;
 
@@ -211,7 +211,7 @@ void HE::InlinePoolAllocatorTest::Prepare()
         }
     });
 
-    AddTest("Perf. & Stress, Size(512)", [&, this](auto &ls) {
+    AddTest("Perf. & Stress, Size(512)", [&, this](auto& ls) {
         constexpr int inlineSize = 512;
         InlinePoolAllocator<int, inlineSize> inlineAlloc;
 
@@ -223,7 +223,7 @@ void HE::InlinePoolAllocatorTest::Prepare()
         }
     });
 
-    AddTest("Perf. & Stress, Size(1024)", [&, this](auto &ls) {
+    AddTest("Perf. & Stress, Size(1024)", [&, this](auto& ls) {
         constexpr int inlineSize = 1024;
         InlinePoolAllocator<int, inlineSize> inlineAlloc;
 
@@ -235,7 +235,7 @@ void HE::InlinePoolAllocatorTest::Prepare()
         }
     });
 
-    AddTest("Perf. & Stress, Size(2048)", [&, this](auto &ls) {
+    AddTest("Perf. & Stress, Size(2048)", [&, this](auto& ls) {
         constexpr int inlineSize = 2048;
         InlinePoolAllocator<int, inlineSize> inlineAlloc;
 
@@ -247,7 +247,7 @@ void HE::InlinePoolAllocatorTest::Prepare()
         }
     });
 
-    AddTest("Perf. & Stress, Size(8192)", [&, this](auto &ls) {
+    AddTest("Perf. & Stress, Size(8192)", [&, this](auto& ls) {
         constexpr int inlineSize = 8192;
         InlinePoolAllocator<int, inlineSize> inlineAlloc;
 
@@ -259,7 +259,7 @@ void HE::InlinePoolAllocatorTest::Prepare()
         }
     });
 
-    AddTest("Vector Growth, size(32)", [&, this](auto &ls) {
+    AddTest("Vector Growth, size(32)", [&, this](auto& ls) {
         constexpr int inlineSize = 32;
         ls << "==========================================" << lf;
         ls << "Inline Alloc Size = " << inlineSize << lf;
@@ -271,7 +271,7 @@ void HE::InlinePoolAllocatorTest::Prepare()
         }
     });
 
-    AddTest("Vector Growth, size(128)", [&, this](auto &ls) {
+    AddTest("Vector Growth, size(128)", [&, this](auto& ls) {
         constexpr int inlineSize = 128;
         ls << "==========================================" << lf;
         ls << "Inline Alloc Size = " << inlineSize << lf;
@@ -283,7 +283,7 @@ void HE::InlinePoolAllocatorTest::Prepare()
         }
     });
 
-    AddTest("Vector Growth, size(512)", [&, this](auto &ls) {
+    AddTest("Vector Growth, size(512)", [&, this](auto& ls) {
         constexpr int inlineSize = 512;
         ls << "==========================================" << lf;
         ls << "Inline Alloc Size = " << inlineSize << lf;
